edje_edit: fix wrong generation of part's source code (effect field).
There is a static array that is used for parsing the current effect
(in Part's struction the effect type is represented by a number) from a number
to text. But there is one type of effects (SHADOW) that has a number combined from
two types. First four bits are always used for defining the effect's type,
but if it is a SHADOW effect, then there is four more bits. They are representing
the direction of the shadow in TEXT block.
This patch fixes code regeneration after saving edje_edit object, so now
it splits the number into two pieces, so we can fully regenerate type of effect.

So, in this path:
- added static array that represents shadow's direction.
- part's source code generaton has been modified.

Reviewers: seoz, raster, cedric

Reviewed By: cedric

CC: reutskiy.v.v, cedric

Differential Revision: https://phab.enlightenment.org/D668

Signed-off-by: Cedric BAIL <cedric.bail@free.fr>